Move ParserError to exceptions.py
